codes clean up for verilator.
