+++ src/main/scala/chiseltest/backends/SimApiInterface.scala
@@ -1,5 +1,5 @@
 // See LICENSE for license details.
-package chiseltest.legacy.backends.verilator
+package chiseltest.backends.verilator
 
 import java.io.File
 import java.nio.channels.FileChannel
@@ -10,18 +10,20 @@ import logger.LazyLogging
 
 import scala.collection.immutable.ListMap
 import scala.collection.mutable
+import scala.collection.mutable.ArrayBuffer
 import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContext, Future, blocking}
 import scala.language.implicitConversions
+import scala.sys.process.{Process, ProcessLogger}
 
 /** Provides an interface layer to an external verilog simulation program through IPC channels
   * Currently supported external programs are those generated by verilator and VCS.
   *
-  * @param cmd  The command to run as a Seq of strings
+  * @param cmd The command to run as a Seq of strings
   */
 private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
-    extends LazyLogging {
-  def getPorts(circuit: Circuit) = circuit.modules.collectFirst { case Module(_, name, ports, _) if name == circuit.main => ports }.get .partition(_.direction == firrtl.ir.Input)
+  extends LazyLogging {
+  def getPorts(circuit: Circuit) = circuit.modules.collectFirst { case Module(_, name, ports, _) if name == circuit.main => ports }.get.partition(_.direction == firrtl.ir.Input)
   //
   // Construct maps for the input and output
   // Remove zero length fields during the process
@@ -49,11 +51,15 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
 
     (ListMap(inputs flatMap genChunk: _*), ListMap(outputs flatMap genChunk: _*))
   }
+
   private object SIM_CMD extends Enumeration {
     val RESET, STEP, UPDATE, POKE, PEEK, FORCE, GETID, GETCHK, FIN = Value
   }
+
   implicit def cmdToId(cmd: SIM_CMD.Value): Int = cmd.id
+
   implicit def int(x: Int): BigInt = (BigInt(x >>> 1) << 1) | BigInt(x & 1)
+
   implicit def int(x: Long): BigInt = (BigInt(x >>> 1) << 1) | BigInt(x & 1)
 
   private var isStale = false
@@ -67,7 +73,9 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
   private[chiseltest] val process = TesterProcess(cmd, _logs)
 
   // Set up a Future to wait for (and signal) the test process exit.
+
   import ExecutionContext.Implicits.global
+
   private[chiseltest] val exitValue = Future(blocking(process.exitValue))
 
   // memory mapped channels
@@ -114,7 +122,7 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
   }
 
   private def throwExceptionIfDead(exitValue: Future[Int]) {
-//    implicit val logger = new TestErrorLog
+    //    implicit val logger = new TestErrorLog
     if (exitValue.isCompleted) {
       val exitCode = Await.result(exitValue, Duration(-1, SECONDS))
       // We assume the error string is the last log entry.
@@ -130,8 +138,8 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
 
   /** A busy-wait loop that monitors exitValue so we don't loop forever if the test application exits for some reason.
     *
-    * @param block  a thunk that determines when complete
-    * @param loop   a thunk to keep running until block is true or exitValue says completed.
+    * @param block a thunk that determines when complete
+    * @param loop  a thunk to keep running until block is true or exitValue says completed.
     */
   private def mwhile(block: => Boolean)(loop: => Unit) {
     while (!exitValue.isCompleted && block) {
@@ -195,7 +203,7 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
       else {
         outChannel.consume()
         Some(
-          ((0 until chunk) foldLeft BigInt(0))(
+          ((0 until chunk) foldLeft BigInt(0)) (
             (res, i) => res | (int(outChannel(i)) << (64 * i))
           )
         )
@@ -211,7 +219,7 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
     if (valid) {
       (outputsNameToChunkSizeMap.toList foldLeft 0) {
         case (off, (out, chunk)) =>
-          _peekMap(out) = ((0 until chunk) foldLeft BigInt(0))(
+          _peekMap(out) = ((0 until chunk) foldLeft BigInt(0)) (
             (res, i) => res | (int(outChannel(off + i)) << (64 * i))
           )
           off + chunk
@@ -228,10 +236,10 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
     if (ready) {
       (inputsNameToChunkSizeMap.toList foldLeft 0) {
         case (off, (in, chunk)) =>
-          val value = _pokeMap getOrElse (in, BigInt(0))
+          val value = _pokeMap getOrElse(in, BigInt(0))
           (0 until chunk) foreach (
             i => inChannel(off + i) = (value >> (64 * i)).toLong
-          )
+            )
           off + chunk
       }
       inChannel.produce()
@@ -308,10 +316,10 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
     mwhile(!recvOutputs) {}
 
     //TODO: Figure out what the behavior of the interface on start up is
-//    for (_ <- 0 until 5) {
-//      mwhile(!sendCmd(SIM_CMD.RESET)) { }
-//      mwhile(!recvOutputs) { }
-//    }
+    //    for (_ <- 0 until 5) {
+    //      mwhile(!sendCmd(SIM_CMD.RESET)) { }
+    //      mwhile(!recvOutputs) { }
+    //    }
   }
 
   def poke(signal: String, value: BigInt) {
@@ -319,9 +327,9 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
       _pokeMap(signal) = value
       isStale = true
     } else {
-      val id = _signalMap getOrElseUpdate (signal, getId(signal))
+      val id = _signalMap getOrElseUpdate(signal, getId(signal))
       if (id >= 0) {
-        poke(id, _chunks getOrElseUpdate (signal, getChunk(id)), value)
+        poke(id, _chunks getOrElseUpdate(signal, getChunk(id)), value)
         isStale = true
       } else {
         logger info s"Can't find $signal in the emulator..."
@@ -342,9 +350,9 @@ private[chiseltest] class SimApiInterface(circuit: Circuit, cmd: Seq[String])
       }
       _pokeMap get signal
     } else {
-      val id = _signalMap getOrElseUpdate (signal, getId(signal))
+      val id = _signalMap getOrElseUpdate(signal, getId(signal))
       if (id >= 0) {
-        Some(peek(id, _chunks getOrElse (signal, getChunk(id))))
+        Some(peek(id, _chunks getOrElse(signal, getChunk(id))))
       } else {
         logger info s"Can't find $signal in the emulator..."
         None
@@ -383,32 +391,76 @@ private[chiseltest] class Channel(name: String) {
   @volatile private lazy val buffer = {
     val size = channel.size
     assert(size > 16, "channel.size is bogus: %d".format(size))
-    channel map (FileChannel.MapMode.READ_WRITE, 0, size)
+    channel map(FileChannel.MapMode.READ_WRITE, 0, size)
   }
+
   implicit def intToByte(i: Int): Byte = i.toByte
+
   val channel_data_offset_64bw = 4 // Offset from start of channel buffer to actual user data in 64bit words.
   def acquire(): Unit = {
-    buffer put (0, 1)
-    buffer put (2, 0)
+    buffer put(0, 1)
+    buffer put(2, 0)
     while ((buffer get 1) == 1 && (buffer get 2) == 0) {}
   }
-  def release(): Unit = { buffer put (0, 0) }
-  def ready: Boolean = (buffer get 3) == 0
-  def valid: Boolean = (buffer get 3) == 1
-  def produce(): Unit = { buffer put (3, 1) }
-  def consume(): Unit = { buffer put (3, 0) }
-  def update(idx: Int, data: Long) {
-    buffer putLong (8 * idx + channel_data_offset_64bw, data)
+
+  def release(): Unit = {
+    buffer put(0, 0)
   }
+
+  def ready: Boolean = (buffer get 3) == 0
+
+  def valid: Boolean = (buffer get 3) == 1
+
+  def produce(): Unit = {
+    buffer put(3, 1)
+  }
+
+  def consume(): Unit = {
+    buffer put(3, 0)
+  }
+
+  def update(idx: Int, data: Long) {
+    buffer putLong(8 * idx + channel_data_offset_64bw, data)
+  }
+
   def update(base: Int, data: String) {
     data.zipWithIndex foreach {
-      case (c, i) => buffer put (base + i + channel_data_offset_64bw, c)
+      case (c, i) => buffer put(base + i + channel_data_offset_64bw, c)
     }
-    buffer put (base + data.length + channel_data_offset_64bw, 0)
+    buffer put(base + data.length + channel_data_offset_64bw, 0)
   }
+
   def apply(idx: Int): Long =
     buffer getLong (8 * idx + channel_data_offset_64bw)
-  def close(): Unit = { file.close() }
+
+  def close(): Unit = {
+    file.close()
+  }
+
   buffer order java.nio.ByteOrder.nativeOrder
   new File(name).delete
 }
+
+
+private[chiseltest] object TesterProcess {
+  def apply(cmd: Seq[String], logs: ArrayBuffer[String]): Process = {
+    require(new java.io.File(cmd.head).exists, s"${cmd.head} doesn't exist")
+    val processBuilder = Process(cmd mkString " ")
+    // This makes everything written to stderr get added as lines to logs
+    val processLogger = ProcessLogger(println, logs += _) // don't log stdout
+    processBuilder run processLogger
+  }
+
+  //TODO: must figure this out before PR
+  def kill(sim: SimApiInterface) {
+    while (!sim.exitValue.isCompleted) sim.process.destroy
+    println("Exit Code: %d".format(sim.process.exitValue))
+  }
+
+  def kill(p: VerilatorBackend[_]) {
+    kill(p.simApiInterface)
+  }
+}
+
+private[chiseltest] case class TestApplicationException(exitVal: Int, lastMessage: String)
+  extends RuntimeException(lastMessage)
