+++ src/main/scala/chiseltest/backends/verilator/Prepare.scala
@@ -1,0 +1,21 @@
+package chiseltest.backends.verilator
+
+import java.io.File
+
+import firrtl.AnnotationSeq
+import firrtl.annotations.NoTargetAnnotation
+import firrtl.options.{Phase, PreservesAll, TargetDirAnnotation}
+
+case class SimulatorHFileDictionary(file: String) extends NoTargetAnnotation
+
+case class SimulatorBinaryPath(file: String) extends NoTargetAnnotation
+
+class Prepare extends Phase with PreservesAll[Phase] {
+  def transform(a: AnnotationSeq): AnnotationSeq = {
+    a :+ a.collectFirst { case f: SimulatorHFileDictionary => f }.getOrElse {
+      SimulatorHFileDictionary(getClass.getResource("/verilator/").getPath)
+    } :+ a.collectFirst { case f: SimulatorBinaryPath => f }.getOrElse {
+      SimulatorBinaryPath("verilator")
+    } :+ a.collectFirst { case TargetDirAnnotation(f) => TargetDirAnnotation(new File(f).getAbsolutePath) }.get
+  }
+}
