+++ src/main/scala/chiseltest/backends/verilator/VerilatorExecutive.scala
@@ -1,4 +1,4 @@
-package chiseltest.legacy.backends.verilator
+package chiseltest.backends.verilator
 
 import chisel3.MultiIOModule
 import chisel3.stage.ChiselCircuitAnnotation
@@ -21,7 +21,7 @@ object VerilatorExecutive extends BackendExecutive {
       Seq(Dependency[chisel3.stage.phases.Elaborate])
     ).transformOrder.foldLeft(chiselAnnotations :+ CompilerAnnotation(new VerilogCompiler()))((a, f) => f.transform(a))
     // run tester2 stage
-    val tester2Annotations = (new chiseltest.stage.VerilatorCompiler).transform(firrtlAnnotations)
+    val tester2Annotations = (new VerilatorCompiler).transform(firrtlAnnotations)
     val dut = getTopModule(chiselAnnotations.collect { case x: ChiselCircuitAnnotation => x }.head.circuit).asInstanceOf[T]
     new VerilatorBackend(dut, tester2Annotations)
   }
