+++ src/main/scala/chiseltest/internal/Testers2.scala
@@ -5,8 +5,7 @@ package chiseltest.internal
 import chisel3.MultiIOModule
 import chiseltest.backends.BackendExecutive
 import chiseltest.backends.treadle.TreadleExecutive
-import chiseltest.legacy.backends.vcs.VcsExecutive
-import chiseltest.legacy.backends.verilator.VerilatorExecutive
+import chiseltest.backends.verilator.VerilatorExecutive
 import firrtl.AnnotationSeq
 import firrtl.annotations.{Annotation, NoTargetAnnotation}
 import firrtl.options.{HasShellOptions, ShellOption, Unserializable}
@@ -71,18 +70,6 @@ case object VerilatorBackendAnnotation extends BackendAnnotation {
       longOption = "t-use-verilator",
       toAnnotationSeq = _ => Seq(VerilatorBackendAnnotation),
       helpText = "direct tester to use verilator backend"
-    )
-  )
-}
-
-case object VcsBackendAnnotation extends BackendAnnotation {
-  val executive: BackendExecutive = VcsExecutive
-
-  val options: Seq[ShellOption[_]] = Seq(
-    new ShellOption[Unit](
-      longOption = "t-use-vcs",
-      toAnnotationSeq = _ => Seq(VerilatorBackendAnnotation),
-      helpText = "direct tester to use VCS backend"
     )
   )
 }
