--- src/main/scala/chiseltest/legacy/backends/vcs/VcsBackend.scala
@@ -1,19 +1,0 @@
-// See LICENSE for license details.
-
-package chiseltest.legacy.backends.vcs
-
-import chisel3._
-import chiseltest.legacy.backends.verilator.VerilatorBackend
-import firrtl.AnnotationSeq
-import firrtl.ir.Circuit
-
-/** Supports Backend and Threaded traits for ex
-  *
-  * @param dut                  the device under test
-  * @tparam T                   the dut's type
-  */
-class VcsBackend[T <: MultiIOModule](
-  dut: T,
-  annos: AnnotationSeq
-) extends VerilatorBackend(dut, annos)
-
